feat: implemented filters for items retrieval endpoint

+++ src/validation/items/companyItems/supplies/createSupplyRequestValidationSchema.ts
@@ -4,7 +4,7 @@ import { z } from "zod";
 export const createSupplyRequestSchema = z
   .object({
     item: z.string().refine((value) => mongoose.Types.ObjectId.isValid(value)),
-    quantity: z.number().min(1),
+    quantity: z.number().min(1).int(),
     price: z.number().min(0.01),
     date: z.coerce.date(),
   })
@@ -12,7 +12,7 @@ export const createSupplyRequestSchema = z
 
 export const updateSupplyRequestSchema = z
   .object({
-    quantity: z.number().min(1),
+    quantity: z.number().min(1).int(),
     price: z.number().min(0.01),
     date: z.coerce.date(),
   })
